+++ starters/apps/basic/src/routes/(starter)/demo/todolist/todolist.module.css
@@ -1,0 +1,44 @@
+.list {
+  display: flex;
+  flex-direction: column;
+  gap: 20px;
+  color: white;
+}
+
+.list,
+.empty {
+  min-height: 250px;
+}
+
+.list li {
+  list-style: none;
+}
+
+.empty {
+  color: white;
+  display: block;
+}
+
+.input {
+  background: white;
+  color: var(--qwik-light-blue);
+  border: none;
+  border-radius: 8px;
+  padding: 15px 20px;
+  margin-right: 10px;
+  font-size: 0.8rem;
+}
+
+.hint {
+  font-size: 0.8rem;
+  color: white;
+  margin-top: 30px;
+}
+
+@media screen and (min-width: 768px) {
+  .input {
+    padding: 23px 35px;
+    margin-right: 20px;
+    font-size: 1rem;
+  }
+}
