--- starters/apps/basic/src/routes/demo/todolist/todolist.module.css
@@ -1,13 +1,0 @@
-.list {
-  display: flex;
-  flex-direction: column;
-  gap: 20px;
-}
-
-.list li {
-  list-style: none;
-}
-
-.hint {
-  font-size: 0.8rem;
-}
